Remove core api and commands relaying on it
- remove core api
- remove marketplace command
- remove ethwallet account command
- move exp:account to account command
