+++ src/commands/account/delete.ts
@@ -1,28 +1,25 @@
-import {WithPassphrase as Command} from '../../account-command'
+import {WithCredential as Command} from '../../credential-command'
 
-export default class AccountDelete extends Command {
+export default class AccountExpDelete extends Command {
   static description = 'Delete an account'
+  static hidden = true
 
   static flags = {
     ...Command.flags,
   }
 
   static args = [{
-    name: 'ADDRESS',
+    name: 'ACCOUNT_NAME',
     required: true
   }]
 
   async run() {
-    const {args} = this.parse(AccountDelete)
-    const passphrase = await this.getPassphrase()
+    const {args} = this.parse(AccountExpDelete)
+    const passphrase = await this.getCredentialPassphrase()
     this.spinner.start('Deleting account')
-    const data = await this.execute({
-      instanceHash: await this.engineServiceInstance(Command.SERVICE_NAME),
-      taskKey: 'delete',
-      inputs: {
-        address: args.ADDRESS,
-        passphrase
-      }
+    const data = await this.api.account.delete({}, {
+      username: args.ACCOUNT_NAME,
+      passphrase: passphrase || '',
     })
     this.spinner.stop()
     return data
