--- src/commands/account/export.ts
@@ -1,31 +1,0 @@
-import {WithPassphrase as Command} from '../../account-command'
-
-export default class AccountExport extends Command {
-  static description = 'Export an existing account'
-
-  static flags = {
-    ...Command.flags,
-  }
-
-  static args = [{
-    name: 'ADDRESS',
-    required: true
-  }]
-
-  async run() {
-    const {args} = this.parse(AccountExport)
-    const passphrase = await this.getPassphrase()
-    this.spinner.start('Exporting account')
-    const data = await this.execute({
-      instanceHash: await this.engineServiceInstance(Command.SERVICE_NAME),
-      taskKey: 'export',
-      inputs: {
-        address: args.ADDRESS,
-        passphrase,
-      }
-    })
-    this.spinner.stop()
-    this.styledJSON(JSON.parse(data.account))
-    return data
-  }
-}
