--- src/commands/account/import-private-key.ts
@@ -1,31 +1,0 @@
-import {WithPassphrase as Command} from '../../account-command'
-
-export default class AccountImportPK extends Command {
-  static description = 'Import an account with a private key'
-
-  static flags = {
-    ...Command.flags,
-  }
-
-  static args = [{
-    name: 'PRIVATE_KEY',
-    description: 'Private key of the account',
-    required: true,
-  }]
-
-  async run() {
-    const {args} = this.parse(AccountImportPK)
-    const passphrase = await this.getPassphrase()
-    this.spinner.start('Importing account')
-    const data = await this.execute({
-      instanceHash: await this.engineServiceInstance(Command.SERVICE_NAME),
-      taskKey: 'importFromPrivateKey',
-      inputs: {
-        privateKey: args.PRIVATE_KEY,
-        passphrase,
-      }
-    })
-    this.spinner.stop(data.address)
-    return data
-  }
-}
