--- src/commands/account/import.ts
@@ -1,31 +1,0 @@
-import {WithPassphrase as Command} from '../../account-command'
-
-export default class AccountImport extends Command {
-  static description = 'Import an account'
-
-  static flags = {
-    ...Command.flags,
-  }
-
-  static args = [{
-    name: 'ACCOUNT',
-    description: 'Account definition in JSON (could be retrieved with account:export)',
-    required: true
-  }]
-
-  async run() {
-    const {args} = this.parse(AccountImport)
-    const passphrase = await this.getPassphrase()
-    this.spinner.start('Importing account')
-    const data = await this.execute({
-      instanceHash: await this.engineServiceInstance(Command.SERVICE_NAME),
-      taskKey: 'import',
-      inputs: {
-        account: args.ACCOUNT,
-        passphrase,
-      }
-    })
-    this.spinner.stop(data.address)
-    return data
-  }
-}
