+++ src/commands/account/list.ts
@@ -2,8 +2,9 @@ import {cli} from 'cli-ux'
 
 import {WithoutPassphrase as Command} from '../../account-command'
 
-export default class AccountList extends Command {
+export default class AccountExpList extends Command {
   static description = 'List accounts'
+  static hidden = true
 
   static flags = {
     ...Command.flags,
@@ -11,14 +12,13 @@ export default class AccountList extends Command {
   }
 
   async run() {
-    const {flags} = this.parse(AccountList)
-    const data = await this.execute({
-      instanceHash: await this.engineServiceInstance(Command.SERVICE_NAME),
-      taskKey: 'list'
-    })
-    cli.table(data.addresses, {
-      address: {header: 'ADDRESS', get: (x: any) => x},
+    const {flags} = this.parse(AccountExpList)
+
+    const {accounts} = await this.api.account.list({})
+    cli.table(accounts, {
+      name: {header: 'NAME', get: x => x.name},
+      address: {header: 'ADDRESS', get: x => x.address},
     }, {printLine: this.log, ...flags})
-    return data
+    return accounts
   }
 }
